requirements.txt
Pillow==8.4.0
torch==1.12.0
torchvision==0.13.0
scikit-image==0.17.2
six==1.16.0
opencv-python==4.6.0.66
statsmodels==0.12.2
numpy==1.19.5
scikit-learn==0.24.2
pandas==1.3.5
scipy==1.5.4
matplotlib==3.5.2
torchdistill==0.3.2
robustness==1.2.1.post2
gco==2.0.2
wand==0.6.8